chore: move inplayer types to its specific integration folder
