+++ packages/common/src/services/integrations/jwp/JWPProfileService.ts
@@ -5,8 +5,8 @@ import ProfileService from '../ProfileService';
 import StorageService from '../../StorageService';
 import type { CreateProfile, DeleteProfile, EnterProfile, GetProfileDetails, ListProfiles, UpdateProfile } from '../../../../types/profiles';
 import { logError } from '../../../logger';
-import type { ProfilesData } from '../../../../types/inplayer';
 
+import type { ProfilesData } from './types';
 import JWPAPIService from './JWPAPIService';
 
 @injectable()
