+++ packages/common/src/services/integrations/jwp/types.ts
